Refactored processing CLI args
Refactored CLI argument processing out of `main` into
a separate function.

This is in preparation for refactoring almost all of
the logic (except for naively taking the CLI args
and handling errors if `lib.rs` returns an error)
and putting the bulk of the logic in `lib.rs`, where
it can be tested.