Changed error to not panic
Refactored `new` to `build` (since it can now fail, and having
a `new` fail is wack) and made it return a `Result`, so that
an error can be handled by `main` instead of giving ugly
programming traces.

Also added correct error handling in `main` so that the
`Err` from `Config::build()` is handled correctly.