Changed error in run to not panic
Similarly to before with `Config::build`, changed `run`
to not panic. Instead, returns a possible error (this time,
any error that implements the `Error` trait) to be handled
by `main`.

Also updated `main` to properly handle it. Unlike with
`Config::build`, just directly do a comparison check
since don't care about the `Ok(())` value that `run`
returns when it passes.