Fixed to output errors to std error
Fixed so that output errors are sent to Standard Error instead
of Standard Output. This comes in important when piping output
to a file (like `cargo run > output.txt`). By having the errors
go to Standard Error, if the program fails the error message
is placed into the terminal, not into the file.